apa gitu wis

+++ database/migrations/2021_11_14_020035_create_pegawais_table.php
@@ -15,7 +15,7 @@ class CreatePegawaisTable extends Migration
     {
         Schema::create('pegawais', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('nik');
+            $table->integer('nik');
             $table->string('nama');
             $table->date('tanggal_lahir');
             $table->string('alamat');
